Move users seed data to mockUsers data object
