+++ db/data/mockUsers.js
@@ -1,0 +1,11 @@
+const users = [
+  {
+    authrocket_id: 'usr_0vYfOixWwPnBDh1w8rxjGm',
+    name: 'Hugh Hartigan',
+    email: 'hartigan.hm@gmail.com',
+    clinic: 'Developmental_FX',
+    clinic_abbreviation: 'DFX'
+  }
+];
+
+export default users;
